+++ html/modules/installer/xaradmin.php
@@ -657,7 +657,7 @@ function installer_admin_create_administrator()
     $data['phase_label'] = xarML('Create Administrator');
 
     sys::import('modules.roles.class.roles');
-    $data['admin'] = xarRoles::getRole(xarModVars::get('roles','admin'));
+    $data['admin'] = xarRoles::getRole((int)xarModVars::get('roles','admin'));
     $data['properties'] = $data['admin']->getProperties();
 
     if (!xarVarFetch('create', 'isset', $create, FALSE, XARVAR_NOT_REQUIRED)) return;
