+++ html/modules/roles/xaradmin/createmail.php
@@ -80,7 +80,7 @@ function roles_admin_createmail()
             $q->eq('r.itemtype',xarRoles::ROLES_USERTYPE);
             $q->ne('r.email','');
         // Set the paging and order stuff for this particular page
-        $numitems = xarModVars::get('roles', 'items_per_page');
+        $numitems = (int)xarModVars::get('roles', 'items_per_page');
         $q->setrowstodo($numitems);
         $q->setstartat($startnum);
         $q->setorder($data['order']);
